+++ pkg/analysis_server/test/lsp/definition_test.dart
@@ -3,11 +3,9 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:analysis_server/lsp_protocol/protocol.dart' as lsp;
-import 'package:analysis_server/src/legacy_analysis_server.dart';
 import 'package:analyzer/src/test_utilities/test_code_format.dart';
 import 'package:analyzer_plugin/protocol/protocol_common.dart';
 import 'package:analyzer_plugin/protocol/protocol_generated.dart' as plugin;
-import 'package:analyzer_testing/experiments/experiments.dart';
 import 'package:test/test.dart';
 import 'package:test_reflective_loader/test_reflective_loader.dart';
 
@@ -22,13 +20,6 @@ void main() {
 
 @reflectiveTest
 class DefinitionTest extends AbstractLspAnalysisServerTest {
-  @override
-  AnalysisServerOptions get serverOptions => AnalysisServerOptions()
-    ..enabledExperiments = [
-      ...super.serverOptions.enabledExperiments,
-      ...experimentsForTests,
-    ];
-
   Future<void> test_acrossFiles() async {
     var mainContents = '''
 import 'referenced.dart';