+++ pkg/analysis_server/test/src/services/correction/fix/remove_dead_code_test.dart
@@ -902,9 +902,6 @@ void f() {
 @reflectiveTest
 class RemoveDeadCodeSoundFlowAnalysisTest extends FixProcessorTest
     with RemoveDeadCodeClassicTestCases {
-  @override
-  List<String> get experiments => [...super.experiments, 'sound-flow-analysis'];
-
   Future<void> test_conditional_unnecessaryNullCheck_equalsNull() async {
     await resolveTestCode(r'''
 class C {