+++ pkg/analyzer/test/src/diagnostics/inference_failure_on_collection_literal_test.dart
@@ -19,7 +19,10 @@ class InferenceFailureOnCollectionLiteralTest extends PubPackageResolutionTest {
   void setUp() {
     super.setUp();
     writeTestPackageAnalysisOptionsFile(
-      analysisOptionsContent(experiments: experiments, strictInference: true),
+      analysisOptionsContent(
+        experimentalFeatures: experimentalFeatures,
+        strictInference: true,
+      ),
     );
   }