+++ pkg/analyzer/test/src/diagnostics/invalid_use_of_internal_member_test.dart
@@ -26,7 +26,7 @@ class InvalidUseOfInternalMemberTest extends PubPackageResolutionTest {
     super.setUp();
     newAnalysisOptionsYamlFile(
       fooPackageRootPath,
-      analysisOptionsContent(experiments: experiments),
+      analysisOptionsContent(experimentalFeatures: experimentalFeatures),
     );
     writeTestPackageConfig(
       PackageConfigFileBuilder()